Minor: Moved unit test class to right place
